cli: better formatting for Help struct
Previously, we'd return the help text no matter what format arguments
were passed to Help.Format(). This made it hard to troubleshoot what
was happening, why we were printing out the same text twice...

This is not a perfect implementation of Help string formatting but it
should improve things slightly. There is a little bit of momentum in
the standard library to standardize improvements to fmt.Formatter and
fmt.State, that we should implement when they are available.