Trim flags parsed from option tag
The short and long flags for an option are expected to be in a
comma-separated list, but spaces used after the commas should be
ignored. When they aren't, then a long flag is not detected as such
because it is parsed as " --value" (and not "--value"), and that parsed
value does not start with "--".

The flag parsing is fixed so that each parsed value is trimmed.

Resolves https://github.com/segmentio/cli/issues/21