human: fix parsing issues
The existing implementation of countPrefix had an off-by-one error
where if the prefix was not found, we would return (len - 1)
instead of returning (len), which led to nonsensical results like
parseFloat("98") returning 9 instead of 98.

Fix that error and also fix duration parsing to return a better error
message if you don't include a unit in the provided duration.