cli,command: fix error representation
If a parse error occurs on a command flag, catch and re-throw the
error with more context in the *CommandFunc.Call function. Otherwise
the only context you get is the failed flag, without a list of all of
the options or what they mean.

Also, only print out Cmd and Err if they are non-nil, otherwise you
get "Command: <nil>" in the output, which is very confusing.